+++ src/features/reader/screens/Reader.tsx
@@ -28,6 +28,7 @@ import { getErrorMessage } from '@/lib/HelperFunctions.ts';
 import type { NavbarContextType } from '@/features/navigation-bar/NavigationBar.types.ts';
 import { withPropsFrom } from '@/base/hoc/withPropsFrom.tsx';
 import { useReaderResetStates } from '@/features/reader/hooks/useReaderResetStates.ts';
+import { useReaderWakeLock } from '@/features/reader/wake-lock/useReaderWakeLock.ts';
 import { useReaderSetSettingsState } from '@/features/reader/hooks/useReaderSetSettingsState.ts';
 import { useReaderShowSettingPreviewOnChange } from '@/features/reader/hooks/useReaderShowSettingPreviewOnChange.ts';
 import { useReaderSetChaptersState } from '@/features/reader/hooks/useReaderSetChaptersState.ts';
@@ -69,6 +70,7 @@ const BaseReader = ({
         backgroundColor,
         shouldShowReadingModePreview,
         shouldShowTapZoneLayoutPreview,
+        shouldWakeLockScreen,
         useAutoBackgroundColorContinuousMode,
     } = useReaderSettingsStore(
         'shouldSkipDupChapters',
@@ -76,6 +78,7 @@ const BaseReader = ({
         'backgroundColor',
         'shouldShowReadingModePreview',
         'shouldShowTapZoneLayoutPreview',
+        'shouldWakeLockScreen',
         'useAutoBackgroundColorContinuousMode',
     );
     const readingMode = useReaderSettingsStore('readingMode');
@@ -157,6 +160,7 @@ const BaseReader = ({
     }, [mangaResponse.data?.manga]);
 
     useReaderResetStates();
+    useReaderWakeLock(shouldWakeLockScreen);
     useReaderSetSettingsState(
         mangaResponse,
         defaultSettingsResponse,